Plus operation spec
The problem is that we should have generic types to implement a proposed code composition. Also we might want to incorporate some kind of traits with strict types, which is extremely undesirable in current design. Perhaps repeating the code in different components would be a great idea(no). So, the general language structure and consistency leave a lot to be desired.