remove N from For component, use repeat macro
I think this is how it should be implemented. Repeat macro takes block of code and just repeats it line by line in file, this kind of expansion could be found in any HDL language, but you usually make this expansion by hand. One question left is how to iterate, what if we won't to iterate throw each second element, what if we want to apply lambda to collection? This will become clearer when will have some draft for catamorphisms and different traversals.