list type
This proposal extends the core of the language and creates new basic type List. List is a managing part of the language which allows to describe some meta information about libraries and general state of platform. It might be nice to describe documentation and test structures on the same level of importance.